currency

+++ app/pages/premiumlease/PremiumsGrantLeasePage.scala
@@ -21,13 +21,13 @@ import play.api.libs.json.JsPath
 
 import scala.language.postfixOps
 
-case object PremiumsGrantLeasePage extends QuestionPage[Int] {
+case object PremiumsGrantLeasePage extends QuestionPage[BigDecimal] {
 
   override def path: JsPath = JsPath \ toString
 
   override def toString: String = "premiumsGrantLease"
 
-  def calculateTaxableAmount(premiumAmount: Int, periods: Int): Int = premiumAmount*((50-minusOne(periods))*50)
+  def calculateTaxableAmount(premiumAmount: BigDecimal, periods: Int): BigDecimal = premiumAmount*((50-minusOne(periods))*50)
 
   def minusOne(periods: Int): Int = periods - 1
 }
